🧩 WN3ZS8 tests: commit split hook files
